Move 'MakeTableWithTruncatedColumn' to asciitable and truncate labels
